BUG FIX: added sort to sha-1 reification
The sha-1 reification did not previously sort the symbols used to form
the string that gets hashed. This lead to the possibility of different
hashes being produced for the same symbols. By sorting prior to
computing the hash, we are now guaranteed to produce the same sha-1
hash given a set of input symbols.